Datos ajustados

+++ src/app/Components/tabla-user/tabla-user.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { UserModel } from 'src/app/model/userModel';
 import { UserService } from 'src/app/Services/user.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-tabla-user',
@@ -28,9 +29,28 @@ export class TablaUserComponent implements OnInit {
     });
   }
   eliminarUser(user: UserModel) {
-    this._serviceUser.eliminarUser(user).subscribe(data => {
-      console.log('elimino');
-      this.listarUser();
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "Desea elminar este registro",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, Eliminar!'
+    }).then((result) => {
+      if (result.value) {
+        Swal.fire(
+          'Deleted!',
+          'Se elimino correctamente',
+          'success'
+        );
+        this._serviceUser.eliminarUser(user).subscribe(data => {
+          console.log('elimino');
+          this.listarUser();
+        });
+
+      }
     });
 
   }
